pyproject.toml
[tool.poetry]
name = "dataprocessing"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.1"
confluent-kafka = "^1.7.0"
requests = "^2.25.1"
avro = "^1.10.2"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
